Correct simplify return type annotation (int32, not int64)
``simplify`` is annotated as returning ``NDArray[np.int64]`` faces and
collapses, but it always returns int32: faces come from
``_simplify.return_faces_int32_no_padding()`` and collapses from
``_simplify.return_collapses()`` (a ``NDArray<int32_t, 2>`` in the core).
This holds regardless of the input face dtype, since both int32 and int64
face inputs are read back through the int32 accessors.

Fix the return annotation to ``NDArray[np.int32]`` for the faces and
collapses entries (points remain ``NDArray[np.float64]``), and extend
``test_simplify_int64_faces_matches_int32`` to assert the returned faces
are int32 for both int32 and int64 inputs.

Annotation-only change; no runtime behavior or C++ is touched.

+++ fast_simplification/simplify.py
@@ -46,8 +46,8 @@ def simplify(
     lossless: bool = False,
     preserve_border: bool = False,
 ) -> (
-    tuple[NDArray[np.float64], NDArray[np.int64]]
-    | tuple[NDArray[np.float64], NDArray[np.int64], NDArray[np.int64]]
+    tuple[NDArray[np.float64], NDArray[np.int32]]
+    | tuple[NDArray[np.float64], NDArray[np.int32], NDArray[np.int32]]
 ):
     """Simplify a triangular mesh.
